🍀⛄️ -> Using the terrain tool to create layouts for the controller [SCD-4 SCD-1]
Notes:
The terrain tool is absolutely amazing! I love it
All of the terrain can be updated, we can move things out of the terrain "paintbox" and into individual gameobjects (as prefabs) if we want them to be interactable in-game, etc. Right now just using some very simple assets but we can customise, maybe create custom materials from other assets

When we get this finished and combined with the spherical scene/terrain generator and screenshot functionality, it will be amazing.